Flesh out some helper progs and formats a bit
Add bmp2chr program and modify convert-map program to read child and
main boards.

The maps have been split out into their own files instead of having them
in map.tmx.  The two files use the dimensions of the boards instead of
the whole playfield to make working with the data easier.  The child
boards use the same size tiles as the main boards in the map editor, but
will probably be twice the size in the game itself.

Some auxiliary files have been included until the map format and mapping
process finalizes.

The convert-map utility does not write out any data yet, nor does it
process the map XML data.  It simply reads the XML and prints the data
it finds.  The specifics of the map data format are still being worked
out.

The bmp2chr utility has been brought in from the last project.  It will
be used until the NES utilities are published in their own project.