Add some map size checks
Added in some map dimension checks. The utility that generates the map
data verifies that all layers are the same size.  If there is a layer
with a different size than the first, the build will fail.  It also
creates two labels with the values of the width and the height that are
checked with an ASSERT in the assembly source.