Make the ball bounce off the walls
Bounce the ball off the walls using a single central collision point on
the ball and a bounding box for the walls.

The bounce takes into consideration the distance into the wall the ball
is on the collision and adds that distance to the ball in the opposite
direction.

The ball can move at 10 pixels per frame without clipping through we
walls.