Rewrite credit data generation
Credit data is changing a lot.  Moving to lists of just names and an
attribution list (label/title + name).  Attribution list isn't
implemented yet.

Subscriber names are grouped by tenure with the same divisions that sub
badges use on Twitch.

Input is now a directory that contains a collection of CSV files.
These CSV files are downloaded manually from Twitch.  The only piece of
data not used from this file is the streak length, everything else is
used in some way.

This data is not currently consumed by the assembly.  This also means
that this change breaks the build.