Read the new credits data format in assembly
Separate ram into global and non-global portions.  The global ram is
just a single page starting at $0300 while the non-global ram starts at
$0400 and continues through $07FC.  Each game state has it's own
non-global ram allocations and this portion of memory should be cleared
when changing the game state.

Modified generate-credits.go to write a static attribution group.  This
will need to be removed in favor of loading this data from a file
instead.

Modified the credits code to load the new data format.  The groups are
not handled differently at this point, and the attribution group is
ignored entirely.