Change the paddle bounce to static values
Instead of adding and subtracting values from the ball speed when it
bounces on the paddle, set it to a static value from a lookup table.
The lookup is still using the distance from the center of the paddle.

The only value not coming from the lookup table is when the ball hits
the paddle dead center, in which case the ball bounces straight up with
no horizontal movement.