Add a boost to the ball
Pressing A will boost the ball vertically by a set amount.  The boost
amount is added to the vertical speed in such a way that will move speed
it up (if traveling up) or slow it down (if traveling down).

The game starts with a boost pool size of two.  Each use reduces the
available boosts by one.  Each bounce on the paddle increases the
available boosts by one up until the max size of the pool.

Also added a sprite to display the number of currently available boosts.