Refactor the credits' frame logic
Refactored the credit screen's frame logic to properly handle single
screen groups as well as scrolling over multi-screen groups.

For a group that only has enough names for a single screen, do not
scroll at all.  Only pause for a little bit before moving on to the next
group.

Groups with enough names for more than one screen will pause before
scrolling.  When the end of the list of names is reached, the screen
will scroll until the last name is on screen then wait before moving on
to the next group.  The amount of wait time should be the same as the
single screen wait.