Use sprite "animation objects"
Instead of duplicating animation data for each individual sprite, an
"animation object" is instead defined with no duplicate data.  The only
data outside the object definion is the base X/Y coordinate for that
instance of the object.

This change makes it easier to define and modify animation data for the
level select screen.