Implement palette animations
A new field was added to the sprite animation objects designating the
animation type.  0 is the original movement animation, while 1 is a
static "animation" that doesn't actually move.

Palette animations have been implemented separately from sprite objects,
but follow a similar method.  A buffer has been added to Zero Page where
all the palettes are expected to be written (animated or otherwise).