Draw a background on the level select screen
The icons have been moved to a background data layer.  They will be
drawn with the rest of the background instead of individually.

This new background is laid out in Tiled and converted to an ASM file by
some new code in the map conversion utility.  It's not finished, but
it's a start.