Fix off-by-one error when getting a brick address
This also fixes GetAddressesForBrick.  Originally the assumption was
that this routine would only be run when a brick was present, however it
was being called before a colission was confirmed.  Corrected the
calling routines to only call GetAddressesForBrick after a collision was
confirmed.

Also added in a bunch of .ifdef's with sanity checks and debugging code.