Imelement DrawTextFromTable (sf_DrawTextFromTable)
This is a scene function used to draw text that is stored in a lookup
table.  The table is currently hard-coded to the label DialogueIndex.

This function sets a flag and just calls DrawText.  DrawText has been
modified to look for this flag, and if set will read the text from the
lookup table instead of the data stream.

SetDialogueTable (sf_SetDialogueTable) has been semi-implemented, but
not added to the function enum.  It will currently just set a pointer,
but nothing is done with that pointer.  If this function is eventually
used, the hard-coded DialogueIndex read will need to be swapped out for
a pointer dereference.

Additional scene data has been added to take advantage of this new
functionality.  This hasn't been tested yet and probably doesn't work.