Implement partial text drawing
Text can now be drawn while the PPU is turned on.  One tile is written
to the PPU per frame.  No new commands are added for this specifically.

Added the ClearText scene command to clear two whole rows of tiles.
This includes the ClearTileRow function in the text engine to implement
clearing the CHR RAM in an unrolled loop.